requirements.txt
setuptools
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
django-environ==0.11.2
python-dotenv==1.0.1
selenium==4.25.0
pgvector
torch
numpy
transformers
llama-index-core>=0.11.13.post1
llama-index-embeddings-huggingface
llama-index-readers-database
llama-index-vector-stores-postgres
langchain
vllm
PyJWT
python-dotenv
django-cors-headers
djangorestframework-simplejwt
langchain-experimental
langchain-openai
spacy
huspacy
cohere
matplotlib>=3.8.0
# https://huggingface.co/huspacy/hu_core_news_lg/resolve/main/hu_core_news_lg-any-py3-none-any.whl
